Separate some of the files in Aggregates folder to be closer to one-type-one-file.
